Refactoring post review 1.
