+++ .gitignore
@@ -1,11 +1,7 @@
 
 -# Compiled class file
 -*.class
- 
- 
--# Log file
 
--
 -# BlueJ files
 -*.ctxt
 -
