+++ daily-patch-information/src/main/java/org/wso2/engineering/efficiency/patch/analysis/jira/JIRAIssue.java
@@ -1,42 +1,40 @@
-//
-// Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-//
-// WSO2 Inc. licenses this file to you under the Apache License,
-// Version 2.0 (the "License"); you may not use this file except
-// in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-//
-package org.wso2.patchinformation.jira;
+/*
+Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+WSO2 Inc. licenses this file to you under the Apache License,
+Version 2.0 (the "License"); you may not use this file except
+in compliance with the License.
+You may obtain a copy of the License at
 
-import org.wso2.patchinformation.constants.Constants;
-import org.wso2.patchinformation.email.HtmlTableRow;
-import org.wso2.patchinformation.pmt.InactivePatch;
-import org.wso2.patchinformation.pmt.OpenPatch;
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied.  See the License for the
+specific language governing permissions and limitations
+under the License.
+*/
+
+package org.wso2.engineering.efficiency.patch.analysis.jira;
+
+import org.wso2.engineering.efficiency.patch.analysis.constants.Constants;
+import org.wso2.engineering.efficiency.patch.analysis.email.HtmlTableRow;
+import org.wso2.engineering.efficiency.patch.analysis.pmt.InactivePatch;
+import org.wso2.engineering.efficiency.patch.analysis.pmt.OpenPatch;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
-
-import static org.wso2.patchinformation.constants.Constants.JIRA_URL_PREFIX;
-import static org.wso2.patchinformation.constants.Constants.NOT_SPECIFIED;
-import static org.wso2.patchinformation.constants.Constants.NO_ENTRY_IN_PMT;
+import java.util.Objects;
 
 /**
  * Represents a JIRA Issue.
  */
-public class JIRAIssue implements HtmlTableRow {
+public class JIRAIssue implements HtmlTableRow, Comparable<JIRAIssue> {
 
     private String name;
+    private String jiraLink;
     private String assignee;
     private String jiraCreateDate;
     private String jiraState;
@@ -45,11 +43,12 @@ public class JIRAIssue implements HtmlTableRow {
     private ArrayList<InactivePatch> inactivePatches;
     private String reportDate; //oldest of all the patches associated with the JIRA
     private String reportDateReleasedPatches; //oldest of all the released patches associated with the JIRA
-    private String jiraLink;
+
 
     JIRAIssue(String jiraName, String assignee, String jiraCreateDate, String jiraState) {
+
         this.name = jiraName;
-        this.jiraLink = JIRA_URL_PREFIX + jiraName;
+        this.jiraLink = Constants.JIRA_URL_PREFIX + jiraName;
         this.assignee = assignee;
         this.jiraCreateDate = stripDate(jiraCreateDate);
         this.jiraState = jiraState;
@@ -57,19 +56,50 @@ public class JIRAIssue implements HtmlTableRow {
         this.openPatches = new ArrayList<>();
         this.releasedPatches = new ArrayList<>();
         this.inactivePatches = new ArrayList<>();
-        this.reportDate = NO_ENTRY_IN_PMT;
-        this.reportDateReleasedPatches = NO_ENTRY_IN_PMT;
+        this.reportDate = Constants.NO_ENTRY_IN_PMT;
+        this.reportDateReleasedPatches = Constants.NO_ENTRY_IN_PMT;
+    }
+
+    /**
+     * Returns the oldest of two date.
+     *
+     * @param currentDateStr the date val currently recorded.
+     * @param newDateStr     new date.
+     * @return the oldest date.
+     */
+    private static String dateCompare(String currentDateStr, String newDateStr) {
+
+        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
+        try {
+            Date currentDate = sdf.parse(currentDateStr);
+            Date newDate = sdf.parse(newDateStr);
+            if (currentDate.before(newDate)) {
+                return currentDateStr;
+            } else {
+                return newDateStr;
+            }
+        } catch (ParseException e) {
+            return currentDateStr;
+        }
     }
 
     public String getJiraState() {
         return jiraState;
     }
 
-    public String getReportDate() {
+    private String getReportDate() {
         return reportDate;
     }
 
-    public String getReportDateReleasedPatches() {
+    public void setReportDate(String currentReportDate) {
+        if (Constants.NO_ENTRY_IN_PMT.equals(reportDate)) {
+            reportDate = currentReportDate;
+        } else {
+            reportDate = dateCompare(reportDate, currentReportDate);
+        }
+    }
+
+    String getReportDateReleasedPatches() {
         return reportDateReleasedPatches;
     }
 
@@ -97,14 +127,6 @@ public class JIRAIssue implements HtmlTableRow {
         return this.assignee;
     }
 
-    public void setReportDate(String currentReportDate) {
-        if (NO_ENTRY_IN_PMT.equals(reportDate)) {
-            reportDate = currentReportDate;
-        } else {
-            reportDate = dateCompare(reportDate, currentReportDate);
-        }
-    }
-
     public void addInactivePatch(InactivePatch patch) {
         this.inactivePatches.add(patch);
     }
@@ -117,8 +139,9 @@ public class JIRAIssue implements HtmlTableRow {
     }
 
     private void addReleasedPatches(OpenPatch patch, String currentReportDate) {
+
         this.releasedPatches.add(patch);
-        if (NO_ENTRY_IN_PMT.equals(reportDateReleasedPatches)) {
+        if (Constants.NO_ENTRY_IN_PMT.equals(reportDateReleasedPatches)) {
             reportDateReleasedPatches = currentReportDate;
         } else {
             reportDateReleasedPatches = dateCompare(reportDateReleasedPatches, currentReportDate);
@@ -132,6 +155,7 @@ public class JIRAIssue implements HtmlTableRow {
      * @return Returns the HTML code for a table row.
      */
     public String objectToHTML(String backgroundColor) {
+
         return "<tr><td width=\"" + "30%" + "\" align=\"left\" bgcolor=" + backgroundColor +
                 " style=\"font-family: Open Sans, Helvetica, Arial, sans-serif; font-size: 14px; font-weight:" +
                 " 400; line-height: 20px; padding: 15px 10px 5px 10px;\">" +
@@ -148,12 +172,13 @@ public class JIRAIssue implements HtmlTableRow {
     }
 
     /**
-     * Builds the JIRA dev patches data as a HTML table row.
+     * Builds the JIRA data on dev patches as a HTML table row.
      *
      * @param backgroundColor of table row.
      * @return Returns the HTML code for a table row.
      */
     public String devPatchesToHTML(String backgroundColor) {
+
         return "<tr><td width=\"" + "30%" + "\" align=\"left\" bgcolor=" + backgroundColor +
                 " style=\"font-family: Open Sans, Helvetica, Arial, sans-serif; font-size: 14px; font-weight:" +
                 " 400; line-height: 20px; padding: 15px 10px 5px 10px;\">" +
@@ -170,29 +195,6 @@ public class JIRAIssue implements HtmlTableRow {
     }
 
     /**
-     * Returns the oldest of two date.
-     *
-     * @param currentDateStr the date val currently recorded.
-     * @param newDateStr     new date.
-     * @return the oldest date.
-     */
-    private static String dateCompare(String currentDateStr, String newDateStr) {
-
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-        try {
-            Date currentDate = sdf.parse(currentDateStr);
-            Date newDate = sdf.parse(newDateStr);
-            if (currentDate.before(newDate)) {
-                return currentDateStr;
-            } else {
-                return newDateStr;
-            }
-        } catch (ParseException e) {
-            return currentDateStr;
-        }
-    }
-
-    /**
      * Returns the date from a "date time" value.
      *
      * @param dateTime date time value returned by JIRA.
@@ -200,12 +202,42 @@ public class JIRAIssue implements HtmlTableRow {
      */
     private String stripDate(String dateTime) {
         if (dateTime == null || !(dateTime.contains("T"))) {
-            return NOT_SPECIFIED;
+            return Constants.NOT_SPECIFIED;
         } else {
             String[] dateSplit = dateTime.split("T");
             return dateSplit[0];
         }
     }
 
+    @Override
+    public int compareTo(JIRAIssue jiraIssue1) {
+        return this.getReportDate().compareTo(jiraIssue1.getReportDate());
+    }
 
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        JIRAIssue jiraIssue = (JIRAIssue) o;
+        return Objects.equals(name, jiraIssue.name) &&
+                Objects.equals(jiraLink, jiraIssue.jiraLink) &&
+                Objects.equals(assignee, jiraIssue.assignee) &&
+                Objects.equals(jiraCreateDate, jiraIssue.jiraCreateDate) &&
+                Objects.equals(jiraState, jiraIssue.jiraState) &&
+                Objects.equals(openPatches, jiraIssue.openPatches) &&
+                Objects.equals(releasedPatches, jiraIssue.releasedPatches) &&
+                Objects.equals(inactivePatches, jiraIssue.inactivePatches) &&
+                Objects.equals(reportDate, jiraIssue.reportDate) &&
+                Objects.equals(reportDateReleasedPatches, jiraIssue.reportDateReleasedPatches);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name, jiraLink, assignee, jiraCreateDate,
+                jiraState, openPatches, releasedPatches, inactivePatches, reportDate, reportDateReleasedPatches);
+    }
 }
