+++ daily-patch-information/src/main/java/org/wso2/engineering/efficiency/patch/analysis/jira/ReportDateComparator.java
@@ -1,0 +1,31 @@
+/*
+Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+WSO2 Inc. licenses this file to you under the Apache License,
+Version 2.0 (the "License"); you may not use this file except
+in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied.  See the License for the
+specific language governing permissions and limitations
+under the License.
+*/
+package org.wso2.engineering.efficiency.patch.analysis.jira;
+
+import java.io.Serializable;
+import java.util.Comparator;
+
+
+/**
+ * Implements the Comparator class to order objects of the JIRAIssue class by it's ReportDateReleasedPatches attribute.
+ */
+public class ReportDateComparator implements Comparator<JIRAIssue>, Serializable {
+
+    public int compare(JIRAIssue jiraIssue1, JIRAIssue jiraIssue2) {
+        return jiraIssue1.getReportDateReleasedPatches().compareTo(jiraIssue2.getReportDateReleasedPatches());
+    }
+}
