--- issue_template.md
@@ -1,18 +1,0 @@
-**Description:**
-<!-- Give a brief description of the issue -->
-
-**Suggested Labels:**
-<!-- Optional comma separated list of suggested labels. Non committers can’t assign labels to issues, so this will help issue creators who are not a committer to suggest possible labels-->
-
-**Suggested Assignees:**
-<!--Optional comma separated list of suggested team members who should attend the issue. Non committers can’t assign issues to assignees, so this will help issue creators who are not a committer to suggest possible assignees-->
-
-**Affected Product Version:**
-
-**OS, DB, other environment details and versions:**    
-
-**Steps to reproduce:**
-
-
-**Related Issues:**
-<!-- Any related issues such as sub tasks, issues reported in other repositories (e.g component repositories), similar problems, etc. -->
